Move test-protozero-trace.cc to toplevel and add it in the autotools build where it was missing
